Сделал сервисы

+++ TicketSelling.Services/Validator/Validators/StaffModelValidator.cs
@@ -25,7 +25,7 @@ namespace TicketSelling.Services.Validator.Validators
                 .NotNull().WithMessage(MessageForValidation.DefaultMessage)
                 .Length(2, 50).WithMessage(MessageForValidation.LengthMessage);
 
-            RuleFor(x => (int)x.Age)
+            RuleFor(x => x.Age)
               .InclusiveBetween(18, 99).WithMessage(MessageForValidation.InclusiveBetweenMessage);
 
             RuleFor(x => x.Post).IsInEnum().WithMessage(MessageForValidation.DefaultMessage);
